remove duplicates in BusStop.getLines
when user clicks on a bus stop it is better to show each line a single time. This is achieved using a Set instead of a List. The set does not allow duplicates. Instead in the opposite direction the BusLine must return a List<BusStop> because the order is important and a line can pass different times at the same stop